rotas egithub api funcionando

+++ server/src/routes/memories.ts
@@ -3,14 +3,20 @@ import { prisma } from '../lib/prisma';
 import { z } from 'zod'
 
 export async function memoriesRoutes(app: FastifyInstance){
-/**METODO LISTA */
+  app.addHook('preHandler', async (request) => {
+    await request.jwtVerify()
+    //garante na requisição se esta vindo o token
+  })
+    
+  /**METODO LISTA */
   app.get('/memories', async(request) =>{ // listagem da memoria
-    await request.jwtVerify() //garante na requisição se esta vindo o token
 
     const memories = await prisma.memory.findMany({
+      where:{
+        userId: request.user.sub,
+      },
       orderBy: {
         createdAt: 'asc', //crescente
-
       },
     })
 
@@ -23,7 +29,7 @@ export async function memoriesRoutes(app: FastifyInstance){
     })
   })
 /**METODO DETALHES */
-  app.get('/memories/:id', async(request) =>{ // detalhes da memoria
+  app.get('/memories/:id', async(request, reply) =>{ // detalhes da memoria
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -34,6 +40,10 @@ export async function memoriesRoutes(app: FastifyInstance){
         id,
       },
     })
+
+    if (!!memory.isPublic && memory.userId !== request.user.sub ){
+      return reply.status(401).send()
+    } // reply é um parametro de envio de dados
 
     return memory
   })
@@ -51,13 +61,13 @@ export async function memoriesRoutes(app: FastifyInstance){
         content,
         isPublic,
         coverUrl,
-        userId: '68349809-1588-4e71-8a74-5536acbb8176',
+        userId: request.user.sub,
       }
     })
     return memory
   })
 /**METODO UPDATE */
-  app.put('/memories/:id', async(request) =>{ // atualizar as memorias
+  app.put('/memories/:id', async(request, reply) =>{ // atualizar as memorias
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
@@ -71,7 +81,17 @@ export async function memoriesRoutes(app: FastifyInstance){
     })
     const { content, isPublic, coverUrl } = bodySchema.parse(request.body)
 
-    const memory = await prisma.memory.update({
+    let memory = await prisma.memory.findUniqueOrThrow({
+      where:{
+        id,
+      }
+    })
+
+    if (memory.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
+
+    memory = await prisma.memory.update({
       where:{
         id,
       },
@@ -84,20 +104,29 @@ export async function memoriesRoutes(app: FastifyInstance){
     return memory
   })
 /**METODO DELETE */
-  app.delete('/memories/:id', async(request) =>{ // apagar memorias
+  app.delete('/memories/:id', async(request, reply) =>{ // apagar memorias
     const paramsSchema = z.object({
       id: z.string().uuid(),
     })
 
     const { id } = paramsSchema.parse(request.params)
-    const memory = await prisma.memory.delete({
+    
+    const memory = await prisma.memory.findUniqueOrThrow({
+      where:{
+        id,
+      }
+    })
+
+    if (memory.userId !== request.user.sub){
+      return reply.status(401).send()
+    }
+    
+    await prisma.memory.delete({
       where: {
         id,
       },
     })
-
-    return memory
   })
 } 
 
-let myTime = 44
+//let myTime = 44
